+++ CHANGELOG.md
@@ -14,6 +14,7 @@ HumHub Changelog
 - Fix #7414: Fix profile field encoding
 - Fix #7419: Reset modal window after close
 - Fix #7428: Fix picker autofocus
+- Enh #7424: New event on get registration groups
 
 1.17.0 (January 13, 2025)
 -------------------------